feat: minor changes in fragments arrange ui components
