+++ app/src/main/java/com/example/calefit/ui/home/select/ExerciseSelectFragment.kt
@@ -1,4 +1,4 @@
-package com.example.calefit.ui.planner
+package com.example.calefit.ui.home.select
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,24 +8,20 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import com.example.calefit.R
 import com.example.calefit.common.autoCleared
-import com.example.calefit.databinding.FragmentPlannerBinding
+import com.example.calefit.databinding.FragmentExerciseSelectBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class PlannerFragment : Fragment() {
+class ExerciseSelectFragment : Fragment() {
 
-    private var binding by autoCleared<FragmentPlannerBinding>()
+    private var binding by autoCleared<FragmentExerciseSelectBinding>()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         binding =
-            DataBindingUtil.inflate(inflater, R.layout.fragment_planner, container, false)
+            DataBindingUtil.inflate(inflater, R.layout.fragment_exercise_select, container, false)
         return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
     }
 }
