+++ app/src/main/java/com/example/calefit/ui/home/template/TemplateFragment.kt
@@ -1,4 +1,4 @@
-package com.example.calefit.ui.select
+package com.example.calefit.ui.home.template
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,20 +8,20 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import com.example.calefit.R
 import com.example.calefit.common.autoCleared
-import com.example.calefit.databinding.FragmentExerciseSelectBinding
+import com.example.calefit.databinding.FragmentTemplateBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class ExerciseSelectFragment : Fragment() {
+class TemplateFragment : Fragment() {
 
-    private var binding by autoCleared<FragmentExerciseSelectBinding>()
+    private var binding by autoCleared<FragmentTemplateBinding>()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         binding =
-            DataBindingUtil.inflate(inflater, R.layout.fragment_exercise_select, container, false)
+            DataBindingUtil.inflate(inflater, R.layout.fragment_template, container, false)
         return binding.root
     }
 }
