+++ app/src/main/res/layout/fragment_exercise_select.xml
@@ -10,20 +10,22 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.select.ExerciseSelectFragment">
+        tools:context=".ui.home.select.ExerciseSelectFragment">
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/til_search"
+            style="@style/Widget.MaterialComponents.TextInputLayout.FilledBox"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginStart="16dp"
-            android:layout_marginTop="16dp"
+            android:layout_marginTop="32dp"
             android:layout_marginEnd="16dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent">
 
             <com.google.android.material.textfield.TextInputEditText
+                style="@style/ExercisePlanBody"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:background="@drawable/edit_text_background"
@@ -32,6 +34,7 @@
         </com.google.android.material.textfield.TextInputLayout>
 
         <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rv_exercise_list"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_marginStart="16dp"
@@ -41,6 +44,21 @@
             app:layout_constraintTop_toBottomOf="@id/til_search"
             tools:listitem="@layout/item_exercise_select" />
 
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/btn_exercise_select"
+            style="@style/ExercisePlanBody"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="16dp"
+            android:layout_marginEnd="16dp"
+            android:background="@drawable/template_background"
+            android:gravity="center|center_vertical"
+            android:padding="5dp"
+            android:text="@string/select_exercise"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/rv_exercise_list" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
