+++ app/src/main/res/layout/item_exercise_plan.xml
@@ -9,14 +9,14 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="180dp">
+        android:layout_height="wrap_content">
 
         <TextView
             android:id="@+id/tv_exercise_title"
             style="@style/ExercisePlanTitle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="24dp"
+            android:layout_marginStart="16dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             tools:text="레그프래스" />
@@ -25,65 +25,63 @@
             android:id="@+id/btn_exercise_close"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:padding="8dp"
+            android:layout_marginEnd="16dp"
+            android:padding="4dp"
             android:src="@drawable/ic_baseline_delete_24"
+            android:tint="@color/dark_gray"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <TextView
-            android:id="@+id/tv_exercise_plan_cycle"
-            style="@style/ExercisePlanBody"
-            android:layout_width="wrap_content"
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rv_exercise_cycle"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="18dp"
-            android:text="@string/exercise_plan_cycle"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_exercise_title" />
-
-        <EditText
-            android:id="@+id/et_exercise_plan_cycle"
-            android:layout_width="100dp"
-            android:layout_height="40dp"
-            android:layout_marginTop="10dp"
-            android:background="@drawable/edit_text_background"
-            android:gravity="end|center_vertical"
-            android:hint="@string/edit_cycle"
-            android:inputType="number"
-            android:lines="1"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="16dp"
+            android:layout_marginEnd="16dp"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/btn_exercise_close" />
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/tv_exercise_title"
+            tools:itemCount="3"
+            tools:listitem="@layout/item_exercise_cycle_setup" />
 
-        <TextView
-            android:id="@+id/tv_exercise_plan_weight"
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/btn_cycle_delete"
             style="@style/ExercisePlanBody"
-            android:layout_width="wrap_content"
+            android:layout_width="180dp"
             android:layout_height="wrap_content"
-            android:layout_marginTop="18dp"
-            android:text="@string/exercise_plan_weight"
+            android:layout_marginTop="16dp"
+            android:background="@drawable/template_background"
+            android:gravity="center|center_vertical"
+            android:padding="5dp"
+            android:text="@string/cycle_delete"
+            app:layout_constraintEnd_toStartOf="@id/btn_cycle_add"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_exercise_plan_cycle" />
+            app:layout_constraintTop_toBottomOf="@id/rv_exercise_cycle" />
 
-        <EditText
-            android:id="@+id/et_exercise_plan_weight"
-            android:layout_width="100dp"
-            android:layout_height="40dp"
-            android:layout_marginTop="10dp"
-            android:background="@drawable/edit_text_background"
-            android:gravity="end|center_vertical"
-            android:hint="@string/edit_weight"
-            android:inputType="number"
-            android:lines="1"
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/btn_cycle_add"
+            style="@style/ExercisePlanBody"
+            android:layout_width="180dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="16dp"
+            android:background="@drawable/template_background"
+            android:gravity="center|center_vertical"
+            android:padding="5dp"
+            android:text="@string/cycle_add"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/et_exercise_plan_cycle" />
+            app:layout_constraintStart_toEndOf="@id/btn_cycle_delete"
+            app:layout_constraintTop_toBottomOf="@id/rv_exercise_cycle" />
 
         <View
             android:layout_width="match_parent"
             android:layout_height="1dp"
+            android:layout_marginTop="16dp"
             android:background="@color/gray2"
-            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toStartOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/btn_cycle_add" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
